+++ disparity_view/__init__.py
@@ -1,8 +1,8 @@
 from .view import view_npy
 from .view import depth_overlay, as_colorimage
 from .depth_to_normal import DepthToNormalMap
-from .reprojection import reproject_from_left_and_disparity
-from .reprojection import gen_right_image, make_animation_gif
+from .cv_reprojection import reproject_from_left_and_disparity
+from .cv_reprojection import gen_right_image, make_animation_gif
 
 from .zed_info import (
     get_width_height_fx_fy_cx_cy,
